+++ templates/api-node-express/src/core/errors/index.ts
@@ -1,3 +1,6 @@
+export { AppError } from "../errors-app-error";
+export * from "./contract-errors";
+export { unAuthorizedError } from "./unauthorized-error";
 export * from "./invalid-param-errors";
 export * from "./missing-param-errors";
 export * from "./server-errors";